% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\name{ansi_style}
\alias{ansi_style}
\title{Add ANSI Style Escape Sequences Codes Around Elements of a Character Vector}
\usage{
ansi_style(txt, style, use.style = ansi_available())
}
\arguments{
\item{txt}{character vector to style; every element of the vector will have
escape sequences added at the beginning and end of the string}

\item{style}{character, either length 1L or same length as \code{txt}, what
style to use, see \code{link{ansi_styles}}}

\item{use.style}{logical(1L) whether to use style or not; provides a
mechanism for turning off styling for systems that do not support it}
}
\value{
character vector, \code{txt} with ansi escape sequences added
}
\description{
Prepends and appends the necessary ANSI sequences to produce the requested
style.  Valid styles are those in \code{names(ansistyle:::ansi.codes)}.
Nesting is \bold{not} supported, except for those ANSI sequences that have
different terminators (e.g. font colors vs. background colors). For a better
and more complete implementation see the \code{crayon} package by Gabor
Csardi.
}

