% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\name{ansi_style_nchar}
\alias{ansi_style_nchar}
\title{Count Characters Accounting For ANSI Escape Sequences}
\usage{
ansi_style_nchar(txt, use.style = getOption("ansistyle.use.style"))
}
\arguments{
\item{txt}{character vector to count characters of}

\item{use.style}{TRUE, FALSE, or NULL, whether we should interpret ANSI
escape sequences as escape sequences, or literally; NULL will auto-detect
ansi support with \code{\link{ansi_available}}}
}
\value{
integer same length as \code{txt}
}
\description{
Behaves just like \code{\link{nchar}} when \code{use.style} is \code{FALSE},
and ignores ANSI escape sequences in counts if it is \code{TRUE}
}

