% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\docType{package}
\name{ansistyle-package}
\alias{ansistyle-package}
\title{Style Console Output With ANSI Escape Sequences}
\description{
Stripped down dependency-less version of the \code{crayon} package by
Gabor Csardi. Modifies character vectors by adding ANSI escape sequences to
render text with style in terminals that support ANSI style escape sequences.
}

