% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\name{strip_ansi}
\alias{contains_ansi}
\alias{strip_ansi}
\title{Check For or Remove ANSI Escape Sequences From a String}
\usage{
strip_ansi(string)

contains_ansi(string)
}
\arguments{
\item{string}{character vector}
}
\value{
input, without ANSI escape sequences for \code{strip_ansi}, TRUE or
  FALSE for \code{contains_ansi}
}
\description{
Uses the \code{ansistyle:::ansi_regex} pattern.
}
\examples{
in.red <- ansi_style("hello\\n", "red", use.style=TRUE)
contains_ansi(in.red)
cat(in.red)
contains_ansi(strip_ansi(in.red))
cat(strip_ansi(in.red))
}

