% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\name{valid_ansi_styles}
\alias{valid_ansi_styles}
\title{Return Available ANSI styles}
\usage{
valid_ansi_styles()
}
\value{
character vector with allowable styles
}
\description{
This runs \code{names(ansistyle:::ansi.codes)}.  If you want to look at the
actual codes used try \code{ansistyle:::ansi.codes)}.  If you access the
internal objects directly keep in mind that those may change in the future,
whereas this function should continue to behave as currently documented.
}

