% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ansistyle.R
\name{ansi_available}
\alias{ansi_available}
\title{Attempt to Detect Whether Terminal Supports ANSI Styles}
\usage{
ansi_available(force = getOption("ansistyle.enabled"))
}
\arguments{
\item{force}{TRUE, FALSE, or NULL; set to NULL to allow auto-detection
(default), or TRUE or FALSE to force this function to return TRUE or FALSE}
}
\value{
TRUE or FALSE
}
\description{
Loosely based on \code{crayon:::has_color}, but with some modifications.  In
particular we test for interactive sessions instead of checking that
\code{stdout} is a terminal to allow running with captured output.
}

